# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: FurinaKit API
  description: |
    FurinaKit 是一个多功能在线工具箱，提供 PDF、图片、音视频、文本处理等 48 个工具的 API。
    
    ## 请求格式
    - **文本工具**：`Content-Type: application/json`
    - **文件工具**：`Content-Type: multipart/form-data`
    
    ## 响应格式
    - **文件下载**：返回二进制文件流（带 Content-Disposition 头）
    - **JSON 文本**：返回 `{ success: true, data: { text: "..." } }`
    - **错误**：返回 `{ success: false, error: { code: "...", message: "..." } }`
  version: 0.1.0
  contact:
    name: FurinaKit
servers:
  - url: /furinakit
    description: Production (with basePath)
...
paths:
...
  /api/dev/timestamp:
    post:
      tags: [dev]
      summary: 时间戳转换
      description: 在 Unix 时间戳和日期之间转换
      operationId: timestamp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                mode:
                  type: string
                  enum: [to-date, from-date]
                  default: to-date
                  description: "转换模式：to-date（时间戳转日期）、from-date（日期转时间戳）"
                value:
                  type: string
                  description: "输入值（时间戳或日期字符串）"
                format:
                  type: string
                  enum: [iso, locale, unix]
                  default: iso
                  description: "输出格式（仅 to-date 模式有效）"
      responses:
        '200':
          $ref: '#/components/responses/TextResult'
...
components:
  responses:
    TextResult:
      description: 操作成功（文本结果）
      content:
        application/json:
          schema:
            type: object
            properties:
              success:
                type: boolean
                example: true
              data:
                type: object
                properties:
                  text:
                    type: string
                    description: 结果文本
...